Changed directory, split files.
