+++ .gitignore
@@ -1,8 +1,6 @@
 # Object files
 *.o
-*.ko
 *.obj
-*.elf
 
 # Precompiled Headers
 *.gch
@@ -10,23 +8,19 @@
 
 # Libraries
 *.lib
-*.a
-*.la
-*.lo
 
-# Shared objects (inc. Windows DLLs)
-*.dll
-*.so
-*.so.*
-*.dylib
-
-# Executables
-*.exe
+# Data
 *.out
 *.app
 *.i*86
 *.x86_64
 *.hex
+*.dat
+*.csv
 
 # Debug files
 *.dSYM/
+
+# Subdirectories
+build
+bin/*
