+++ README.md
@@ -2,4 +2,7 @@
 
 Simple gauge theories on the [lattice](https://xkcd.com/505/).
 
+## details
+
 following original `roots' of lattice qcd: [Mike Creutz et al, PRL 1979](http://thy.phy.bnl.gov/~creutz/mypubs/pub031.pdf)
+
