+++ spike/Zn-3d.c
